manual tests (meh)
